authentication-service/pyproject.toml
[tool.poetry]
name = "authentication-service"
version = "0.1.0"
description = ""
authors = ["ov3rwrite <ilyabelykh123@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
pydantic = "^2.6.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
